Refactored: Codebase to be OOP and PSR-4 compliant Removed: Deprecated plugin version header Updated: Tests to use new Utils\Options class Updated: Readme version
