+++ composer.json
@@ -15,5 +15,10 @@
         "phpunit/phpunit": "^6.0",
         "squizlabs/php_codesniffer": "^3.0",
         "wp-coding-standards/wpcs": "^1.0"
+    },
+    "autoload": {
+      "psr-4": {
+          "Mautic\\WP_Mautic\\": "includes/classes/"
+      }
     }
 }
