+++ tests/OptionsTest.php
@@ -3,6 +3,8 @@
  * @package wpmautic\tests
  */
 
+use Mautic\WP_Mautic\Utils\Options;
+
 /**
  * Test Mautic options management
  */
@@ -13,7 +15,7 @@ class OptionsTest extends WP_UnitTestCase
      */
     public function test_invalid_option_name()
     {
-        wpmautic_option('azerty123456');
+        Options::get('azerty123456');
     }
 
     public function test_base_url_when_empty()
@@ -21,7 +23,7 @@ class OptionsTest extends WP_UnitTestCase
         update_option('wpmautic_options', array(
             'base_url' => ''
         ));
-        $this->assertEmpty(wpmautic_option('base_url'));
+        $this->assertEmpty(Options::get('base_url'));
     }
 
     public function test_base_url_with_value()
@@ -29,13 +31,13 @@ class OptionsTest extends WP_UnitTestCase
         update_option('wpmautic_options', array(
             'base_url' => 'http://example.com'
         ));
-        $this->assertEquals('http://example.com', wpmautic_option('base_url'));
+        $this->assertEquals('http://example.com', Options::get('base_url'));
     }
 
     public function test_script_location_when_empty()
     {
         update_option('wpmautic_options', array());
-        $this->assertEquals('header', wpmautic_option('script_location'));
+        $this->assertEquals('header', Options::get('script_location'));
     }
 
     public function test_script_location_with_value()
@@ -43,7 +45,7 @@ class OptionsTest extends WP_UnitTestCase
         update_option('wpmautic_options', array(
             'script_location' => 'footer'
         ));
-        $this->assertEquals('footer', wpmautic_option('script_location'));
+        $this->assertEquals('footer', Options::get('script_location'));
     }
 
     public function test_script_location_with_disabled_value()
@@ -51,13 +53,13 @@ class OptionsTest extends WP_UnitTestCase
         update_option('wpmautic_options', array(
             'script_location' => 'disabled'
         ));
-        $this->assertEquals('disabled', wpmautic_option('script_location'));
+        $this->assertEquals('disabled', Options::get('script_location'));
     }
 
     public function test_fallback_activated_when_empty()
     {
         update_option('wpmautic_options', array());
-        $this->assertTrue(wpmautic_option('fallback_activated'));
+        $this->assertTrue(Options::get('fallback_activated'));
     }
 
     public function test_fallback_activated_with_value()
@@ -65,6 +67,6 @@ class OptionsTest extends WP_UnitTestCase
         update_option('wpmautic_options', array(
             'fallback_activated' => false
         ));
-        $this->assertFalse(wpmautic_option('fallback_activated'));
+        $this->assertFalse(Options::get('fallback_activated'));
     }
 }
